Models a "Game" which records both a board and status
The main win here is that the game itself knows which player
should have the next turn. This _should_ help enforce that player
X can't play twice in a row, for example.

When I do eventually get to the web, I think the server's
responsibility would be to maintain a list of games against
an ID for the game. Then clients can request to make a "move"
against the game's ID and the server can easily verify if the
player making the request can actually make the move.

Two additional things I'd like to explore:

1. Enforce that a PlayerPosition can't be "overwritten".
   If X is in position One then O can't play position One.
2. Model that a game is either in progress or complete