Tries to do all cell validation up front
We'll see if this makes sense - I've been down this road before.

Pros:

- All validation logic happens at once
- The actual playMove function can never fail
- Can probably remove a data type for Game stuff?

Cons:

- Gonna make the tests harder I think?
- Maybe opens up race conditions?
  - could a move be "valid" at a point in time but then when
    actually applied it isn't?
  - probably need to return a the board itself with the move?

We'll see. Maybe what I really want to do is make a version of playMove
that can take a string? I think if at least if all the validation
was in one place it would feel better.